pyproject.toml
[tool.poetry]
name = "jupychat"
version = "0.1.0"
description = "A simple ChatGPT plugin to run jupyter notebooks"
authors = ["Elijah Wilson <dev.tizz98@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.24.1"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
jinja2 = "^3.1.2"
structlog = "^23.1.0"
pyyaml = "^6.0"
pyjwt = "^2.7.0"
aiofiles = "^23.1.0"
nbformat = "^5.8.0"
jupyter-client = "^8.2.0"
ipython = "^8.13.2"
kernel-sidecar = "^0.5.5"
cryptography = "^40.0.2"
ipykernel = "^6.23.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.270"
rich = "^13.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.isort]
line_length = 100
profile = "black"
